backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
insightface==0.7.3
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
numpy==1.24.4
Pillow==10.2.0
pydantic==2.6.1
python-dotenv==1.0.1
qrcode[pil]==7.4.2